refactor: introduce core library
