+++ LogManager.cs
@@ -128,19 +128,4 @@ namespace NodaStack
         }
     }
 
-    public class LogEntry
-    {
-        public DateTime Timestamp { get; set; }
-        public string Message { get; set; } = "";
-        public LogLevel Level { get; set; }
-        public string Service { get; set; } = "";
-    }
-
-    public enum LogLevel
-    {
-        Debug,
-        Info,
-        Warning,
-        Error
-    }
 }
