+++ NodaStack.csproj
@@ -6,10 +6,17 @@
     <Nullable>enable</Nullable>
     <UseWPF>true</UseWPF>
     <UseWindowsForms>true</UseWindowsForms>
+    <EnableWindowsTargeting>true</EnableWindowsTargeting>
+    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
+    <GenerateTargetFrameworkAttribute>false</GenerateTargetFrameworkAttribute>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="System.Text.Json" Version="9.0.0" />
+    <ProjectReference Include="Core/NodaStack.Core.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Remove="Core\**\*.cs" />
   </ItemGroup>
 
 </Project>
