+++ NodaStack.sln
@@ -4,6 +4,8 @@ VisualStudioVersion = 17.5.2.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NodaStack", "NodaStack.csproj", "{DE3518AE-36A6-C8D4-6033-1B2015F76D65}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NodaStack.Core", "Core/NodaStack.Core.csproj", "{E60AC528-85C2-4E11-8EAE-6F41260F7EF6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -12,9 +14,13 @@ Global
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{DE3518AE-36A6-C8D4-6033-1B2015F76D65}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DE3518AE-36A6-C8D4-6033-1B2015F76D65}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DE3518AE-36A6-C8D4-6033-1B2015F76D65}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DE3518AE-36A6-C8D4-6033-1B2015F76D65}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
+               {DE3518AE-36A6-C8D4-6033-1B2015F76D65}.Release|Any CPU.ActiveCfg = Release|Any CPU
+               {DE3518AE-36A6-C8D4-6033-1B2015F76D65}.Release|Any CPU.Build.0 = Release|Any CPU
+               {E60AC528-85C2-4E11-8EAE-6F41260F7EF6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+               {E60AC528-85C2-4E11-8EAE-6F41260F7EF6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+               {E60AC528-85C2-4E11-8EAE-6F41260F7EF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+               {E60AC528-85C2-4E11-8EAE-6F41260F7EF6}.Release|Any CPU.Build.0 = Release|Any CPU
+       EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
