+++ ProjectManager.cs
@@ -235,14 +235,4 @@ echo ""<p>PHP Version: "" . PHP_VERSION . ""</p>"";
         }
     }
 
-    public class ProjectInfo
-    {
-        public string Name { get; set; } = string.Empty;
-        public string Path { get; set; } = string.Empty;
-        public bool HasIndexFile { get; set; }
-        public int FileCount { get; set; }
-        public DateTime LastModified { get; set; }
-        public string ApacheUrl { get; set; } = string.Empty;
-        public string PhpUrl { get; set; } = string.Empty;
-    }
 }
